Refactor an endpoints of the web based editor
