+++ WelsonJS.Toolkit/WelsonJS.Launcher/MainForm.cs
@@ -5,7 +5,6 @@ using System.IO.Compression;
 using System.Security.Principal;
 using System.Threading.Tasks;
 using System.Windows.Forms;
-using WelsonJS.Launcher.Tools;
 
 namespace WelsonJS.Launcher
 {
