+++ WelsonJS.Toolkit/WelsonJS.Launcher/Program.cs
@@ -5,7 +5,6 @@ using System.Linq;
 using System.Threading;
 using System.Windows.Forms;
 using System.Configuration;
-using WelsonJS.Launcher.Tools;
 
 namespace WelsonJS.Launcher
 {
