+++ WelsonJS.Toolkit/WelsonJS.Launcher/app.config
@@ -4,6 +4,7 @@
     <add key="ResourceServerPrefix" value="http://localhost:3000/"/>
     <add key="RepositoryUrl" value="https://github.com/gnh1201/welsonjs"/>
     <add key="CopilotUrl" value="https://copilot.microsoft.com/"/>
+    <add key="DevToolsPrefix" value="http://localhost:9222/"/>
     <add key="AzureAiServiceUrl" value="https://ai-catswords656881030318.services.ai.azure.com/models/chat/completions?api-version=2024-05-01-preview"/>
     <add key="AzureAiServiceApiKey" value=""/>
   </appSettings>
